+++ tests/integration/security/authn/permissive/authn_permissive_test.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 // Package basic contains an example test suite for showcase purposes.
-package security
+package permissive
 
 import (
 	"reflect"
