+++ tests/integration/security/reachability/testdata/global-mtls-off.yaml
