+++ tests/integration/security/reachability/testdata/global-mtls-permissive.yaml
