+++ tests/integration/security/sds_citadel_flow/sds_citadel_flow_test.go
@@ -15,7 +15,6 @@
 package sds_citadel_test
 
 import (
-	"path"
 	"testing"
 	"time"
 
@@ -26,40 +25,13 @@ import (
 	"istio.io/istio/pkg/test/framework/components/istio"
 	"istio.io/istio/pkg/test/framework/components/pilot"
 	"istio.io/istio/pkg/test/util/connection"
+	"istio.io/istio/pkg/test/util/policy"
 	"istio.io/istio/pkg/test/util/retry"
 )
 
 var (
 	inst istio.Instance
 )
-
-type testPolicy struct {
-	t         *testing.T
-	env       *kube.Environment
-	namespace string
-	fileName  string
-}
-
-func (p testPolicy) tearDown() {
-	p.t.Logf("Tearing down policy %q.", p.fileName)
-	if err := p.env.Delete(p.namespace, path.Join("testdata", p.fileName)); err != nil {
-		p.t.Fatalf("Cannot delete %q from namespace %q: %v", p.fileName, p.namespace, err)
-	}
-}
-
-func applyPolicyFile(t *testing.T, env *kube.Environment, namespace string, fileName string) *testPolicy {
-	t.Logf("Applying policy file %v", fileName)
-	if err := env.Apply(namespace, path.Join("testdata", fileName)); err != nil {
-		t.Fatalf("Cannot apply %q to namespace %q: %v", fileName, namespace, err)
-		return nil
-	}
-	return &testPolicy{
-		t:         t,
-		env:       env,
-		namespace: namespace,
-		fileName:  fileName,
-	}
-}
 
 func TestSdsCitadelCaFlow(t *testing.T) {
 	ctx := framework.NewContext(t)
@@ -90,7 +62,7 @@ func TestSdsCitadelCaFlow(t *testing.T) {
 	namespace := istioCfg.SystemNamespace
 	configFile := "global-mtls.yaml"
 
-	policy := applyPolicyFile(t, env, namespace, configFile)
+	policy := policy.ApplyPolicyFile(t, env, namespace, configFile)
 	// Sleep 3 seconds for the policy to take effect.
 	time.Sleep(3 * time.Second)
 	for _, conn := range connections {
@@ -98,7 +70,7 @@ func TestSdsCitadelCaFlow(t *testing.T) {
 			return connection.CheckConnection(t, conn)
 		}, retry.Delay(time.Second), retry.Timeout(10*time.Second))
 	}
-	policy.tearDown()
+	policy.TearDown()
 }
 
 func setupConfig(cfg *istio.Config) {
