+++ tests/integration/security/sds_vault_flow/sds_vault_flow_test.go
@@ -15,7 +15,6 @@
 package sds_vault_test
 
 import (
-	"path"
 	"testing"
 	"time"
 
@@ -26,6 +25,7 @@ import (
 	"istio.io/istio/pkg/test/framework/components/istio"
 	"istio.io/istio/pkg/test/framework/components/pilot"
 	"istio.io/istio/pkg/test/util/connection"
+	"istio.io/istio/pkg/test/util/policy"
 	"istio.io/istio/pkg/test/util/retry"
 )
 
@@ -51,34 +51,6 @@ const (
 		"Wm9\\\\ngqpveVq2olloNbnLNmM3V6F9mqSZACgADmRqf42bixeHczkTfRDKThJcpY5" +
 		"U44vy\\\\nw4Nm32yDWhD6AC68rDkXX68m\\\\n-----END CERTIFICATE-----"
 )
-
-type testPolicy struct {
-	t         *testing.T
-	env       *kube.Environment
-	namespace string
-	fileName  string
-}
-
-func (p testPolicy) tearDown() {
-	p.t.Logf("Tearing down policy %q.", p.fileName)
-	if err := p.env.Delete(p.namespace, path.Join("testdata", p.fileName)); err != nil {
-		p.t.Fatalf("Cannot delete %q from namespace %q: %v", p.fileName, p.namespace, err)
-	}
-}
-
-func applyPolicyFile(t *testing.T, env *kube.Environment, namespace string, fileName string) *testPolicy {
-	t.Logf("Applying policy file %v", fileName)
-	if err := env.Apply(namespace, path.Join("testdata", fileName)); err != nil {
-		t.Fatalf("Cannot apply %q to namespace %q: %v", fileName, namespace, err)
-		return nil
-	}
-	return &testPolicy{
-		t:         t,
-		env:       env,
-		namespace: namespace,
-		fileName:  fileName,
-	}
-}
 
 func TestSdsVaultCaFlow(t *testing.T) {
 	ctx := framework.NewContext(t)
@@ -109,7 +81,7 @@ func TestSdsVaultCaFlow(t *testing.T) {
 	namespace := istioCfg.SystemNamespace
 	configFile := "global-mtls.yaml"
 
-	policy := applyPolicyFile(t, env, namespace, configFile)
+	policy := policy.ApplyPolicyFile(t, env, namespace, configFile)
 	// Sleep 3 seconds for the policy to take effect.
 	time.Sleep(3 * time.Second)
 	for _, conn := range connections {
@@ -117,7 +89,7 @@ func TestSdsVaultCaFlow(t *testing.T) {
 			return connection.CheckConnection(t, conn)
 		}, retry.Delay(time.Second), retry.Timeout(10*time.Second))
 	}
-	policy.tearDown()
+	policy.TearDown()
 }
 
 func setupConfig(cfg *istio.Config) {
